feat(lib): handle default values with transformToFormGroup
BREAKING CHANGE: transformToFormGroup can now receive a `null` value.

It also has a new param "defaultValues" which is a Partial of the form interface.

This gives finer control on the behavior when a null value is passed to a sub form and let you implement what you want. Indeed in some cases you may still display a sub form even if it's value is null, in that case, you then have the choice to reset the formGroup to default values, or set all of them to null. From `transformToFormGroup`, if you return null all the properties of the form will be reset to null, otherwise return an object representing all the values.