fix(lib): remap and root form components don't have the default values passed (only a type issue)
BREAKING CHANGE: Adding a parameter to the `transformToFormGroup` method could be a breaking change if you're calling `super.transformToFormGroup`. As this method is a callback for ngx-sub-form, chances are: You're not calling it and this change will probably **NOT** affect you. The upgrade should go smoothly. If it does affect you, it'll be caught at build time so it should be easy to find out.