feat: rewrite of ngx-sub-form without inheritance
This is a major architecture change which is brought without any breaking change :smile:!

We've split up the code base in 2: Old one and new one.
The old one hasn't moved at all but is now deprecated (not removed yet!).
You can keep using the old one for a bit and have a smooth/incremental update to use the new API.

Few changes that you have to note with the new API:
- Only works with Angular 9 or more
- The app needs to have Ivy activated (this is because we use `ɵmarkDirty` internally. If it ever gets removed we'll probably have to ask to provide the `ChangeDetectorRef` but we were able to around this for now!)
- We got rid of inheritance :raised_hands:
- Form errors on a FormArray are now an object instead of an array. Previously the array contained null values on all the fields without any error. It's now an object containing only the ones with errors and you can access them using the index

Please start upgrading to the new API as soon as possible as we stop supporting the old API as of today and will remove it in a near release.

This closes https://github.com/cloudnc/ngx-sub-form/issues/171 for the major architectural changes and also the following issues as a result:
- closes cloudnc/ngx-sub-form#82
- closes cloudnc/ngx-sub-form#86
- closes cloudnc/ngx-sub-form#93
- closes cloudnc/ngx-sub-form#133
- closes cloudnc/ngx-sub-form#143
- closes cloudnc/ngx-sub-form#144
- closes cloudnc/ngx-sub-form#149
- closes cloudnc/ngx-sub-form#160
- closes cloudnc/ngx-sub-form#168